subscribers/log_processor/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-servicebus
azure-identity
azure-keyvault-secrets
azure-cosmos

